Move helper methods from SlashCommandsBuilder to SlashUtils
